requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pillow>=10.0.0
python-multipart>=0.0.12
numpy>=1.26.0
opencv-python-headless
torch
transformers
